fix: paddr to 34 bit (UInt64)

+++ Sources/SwiftVCore/Bus.swift
@@ -6,7 +6,7 @@ public class Bus {
         devices.append(device)
     }
 
-    func read8(addr: UInt32) -> UInt8 {
+    func read8(addr: UInt64) -> UInt8 {
         switch addr {
         case dram.startAddr...dram.endAddr:
             return dram.read8(addr: addr)
@@ -21,7 +21,7 @@ public class Bus {
         }
     }
 
-    func write8(addr: UInt32, data: UInt8) {
+    func write8(addr: UInt64, data: UInt8) {
         switch addr {
         case dram.startAddr...dram.endAddr:
             dram.write8(addr: addr, data: data)
@@ -37,7 +37,7 @@ public class Bus {
         }
     }
 
-    func read16(addr: UInt32) -> UInt16 {
+    func read16(addr: UInt64) -> UInt16 {
         switch addr {
         case dram.startAddr...dram.endAddr:
             return dram.read16(addr: addr)
@@ -52,7 +52,7 @@ public class Bus {
         }
     }
 
-    func write16(addr: UInt32, data: UInt16) {
+    func write16(addr: UInt64, data: UInt16) {
         switch addr {
         case dram.startAddr...dram.endAddr:
             dram.write16(addr: addr, data: data)
@@ -68,7 +68,7 @@ public class Bus {
         }
     }
 
-    func read32(addr: UInt32) -> UInt32 {
+    func read32(addr: UInt64) -> UInt32 {
         switch addr {
         case dram.startAddr...dram.endAddr:
             return dram.read32(addr: addr)
@@ -83,7 +83,7 @@ public class Bus {
         }
     }
 
-    func write32(addr: UInt32, data: UInt32) {
+    func write32(addr: UInt64, data: UInt32) {
         switch addr {
         case dram.startAddr...dram.endAddr:
             dram.write32(addr: addr, data: data)
